Minimal reproducing example
I want to be able to "go to definition on type "Value" in Messages_t.rs
line 15 (it's a type defined earlier in the file), but it doesn't work.

If I change the name of the module from "spec" to "foo", this the
Go to Definition function then works.

Something that I still don't understand though is that it still can't
find a definition for the verus! macro.